requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
PyJWT==1.7.1
python-dotenv==0.15.0
